+++ src/pages/users/[uid]/lists/[listId]/add-list/index.page.tsx
@@ -4,7 +4,7 @@ import firebase from 'firebase/app';
 import React from 'react';
 import { Headline } from '../../../../../../components/atoms/Headline';
 import Layout from '../../../../../../components/templates/layout';
-import { useAddList } from '../../../../../../hooks/useAddList';
+import { useAddList } from './useAddList';
 
 type AddListProps = {
   postId: string;
